Rename root package to `.all` and add dependencies.
The root package cannot be named `BL_Python` because PyPI deems it "too similar" to some other unknown package.
As a compromise, the package name will now be `BL_Python.all`, but the content is still under the `BL_Python` namespace.

In addition, because the root package should install everything, I have added the other packages as dependencies.