Add GitHub Actions workflow to rewrite BL_Python dependencies during CICD.
Because some BL_Python packages rely on other BL_Python packages in
this same monorepo, changes across packages in a single branch
are likely to cause breakages. This runs a script that will replace
the BL_Python dependency strings in pyproject.toml with strings
that point to the local filesystem. This way, all changes checked
out with the change that triggered the CICD run will be present.

By default, the workflow will rewrite the dependencies. The cases
where the dependencies will not be rewritten are:
- When a tag is created
- When there is no GITHUB_REF envvar set
- When REWRITE_DEPENDENCIES is false for a `workflow_dispatch` event.

Related #33