Separate some aspects of Connexion and Flask initialization.
This is necessary because Connexion 3.x abstracts away Flask to the
point that you don't really interact with Flask directly when using
Connexion. This is unlike the previous version we were using. This
change means many of the Flask APIs we rely on no longer with with
Connexion, and so must be updated to support Connexion 3.x's way of
doing things _in addition to_ Flask's way of doing things.

Related #33