Add a "test" module to the scaffolder.
Also added the ability for modules to control their own rendering. This
allows for looping over templates without having to code it into the
scaffolder itself. The main application scaffolding code can be changed
to support this later on as well. The database module should be updated
to put each table into its own file instead of in __init__.py.

This change also corrects the application hostname used by the templates.