Add `bl-alembic` command.
This command wraps `alembic` and generated a default `alembic.ini` if
one is not specified on the command line. This will allow us to provide
a default configuration file without requiring additional setup by users
of the tool.

Still to figure out: what the convention is for the location of database
migration files in general, and for scaffolded applications. They will
likely not follow the same conventions.