Resolve pyright error in SSM library.
Fixes `ssm.py:132:36 - error: Argument expression after ** must be a mapping with a "str" key type (reportCallIssue)`.

This is caused by attempted to unpack `**None`, which is a runtime
error. Instead, check for `None` and bail out before the undesired
runtime error.

In either case, the behavior is the same - either Python will show the
mapping error (before this fix) or it will show the "not loaded" error
(after) this fix. The behavior of the SSM library itself is unchanged,
so this is only to show a less disturbing message.

Related #92