Change how notifications work for feature flags.
Due to an oversight in what Python allows, the notification methods for
feature flag changes was made abstract, requiring subclasses to
implement it. However, this was intended to be optional because the
feature flag base class is meant to only require implementations for the
get/set interface. To resolve this, the `abstractmethod` decoration has
been removed.

Another change here regards nonsensical notifications in
`CachingFeatureFlagRouter`. The previous notification logged any time
`set` was called, and always said that the flag changed from its
current value to its current value (that is not a typo). In addition
to fixing this, the notification now indicates:

* When an attempt was made to set a flag to its current value
* When a flag was set, and what its new value is
* When a new flag is set and cached

Related #98