Prevent crash when Swagger URL is `/`
When the Swagger URL is `/` or ``, the application crashes on any
request. This is because Connxeion registers conflicting Blueprint URLs
for anything in the OpenAPI spec by using the same base URL as the Swagger
URL. This is not resolved by changing the base path in the OpenAPI spec,
because Connexion will use that base path instead.

Also removed some obsolete and unused code.