Add parameter to selectively mark a release as the "latest."
This change also generates release notes and the title automatically
from the tag. This is something done by default currently done manually,
so this only automates it.
These workflow changes can be used reliably to create a tag and release
through GitHub without needing to explicitly create a tag and push it.